requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pillow
GPUtil
pdfminer.six
easyocr
torch
transformers
huggingface_hub
sentence-transformers
GPUtil
loguru
tqdm
mlflow
prometheus-client
python-jose[cryptography]
passlib[bcrypt]
flupy
gliner